refactor: schema structure
